% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exonCounts}
\alias{exonCounts}
\title{Subset of Pasilla Dataset}
\format{
A data frame with 1000 rows and 7 variables
}
\source{
Exoncounts from Pasilla package \url{https://bioconductor.org/packages/release/data/experiment/html/pasilla.html}
}
\usage{
data("exonCounts")
}
\description{
exonCounts consists of a subset of the exon counts from the pasilla dataset.
}
\references{
Huber W, Reyes A (2020). pasilla: Data package with 
per-exon and per-gene read counts of RNA-seq samples of Pasilla
knock-down by Brooks et al., Genome Research 2011
}
\keyword{datasets}
