\name{origCounts}
\alias{origCounts}
\docType{data}
\title{ Subset of Pasilla Dataset
}
\description{
origCounts consists of a subset of the exon counts from the pasilla dataset.
}
\usage{data("origCounts")}
\format{
A data frame with 1000 observations on the following 7 variables.
\describe{
\item{\code{treated1fb}}{a numeric vector}
\item{\code{treated2fb}}{a numeric vector}
\item{\code{treated3fb}}{a numeric vector}
\item{\code{untreated1fb}}{a numeric vector}
\item{\code{untreated2fb}}{a numeric vector}
\item{\code{untreated3fb}}{a numeric vector}
\item{\code{untreated4fb}}{a numeric vector}
  }
}
\value{
{Dataset}
}
\seealso{\code{\link{EBSEA}}}
\examples{
data(origCounts)
}
\keyword{datasets}
