% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCounts.R
\name{filterCounts}
\alias{filterCounts}
\title{Filter Count Data}
\usage{
filterCounts(x, mean = 1, exonCount = 1)
}
\arguments{
\item{x}{A numeric dataframe of exon counts across the samples.
Exon number in format GeneName:Exonnumber should be indicated
in the row name and sample names as column names.}

\item{mean}{Exons with average count value across the
dataset less than mean are filtered out. Default: 1}

\item{exonCount}{After filtering the individual exons,
only genes with at least the given number of exons
remaining will be retained. Default: 1}
}
\value{
A dataframe of filtered counts of exons
}
\description{
Filtering of exons based on their expression levels
}
\examples{
data(exonCounts)
res <- filterCounts(exonCounts)
}
